Fixed castings

+++ src/generator_gemm_amx.c
@@ -170,7 +170,7 @@ LIBXSMM_API_INTERN void libxsmm_generator_gemm_decompress_KxM_i4_tensor( libxsmm
           LIBXSMM_X86_INSTR_PREFETCHT0,
           i_gp_reg,
           LIBXSMM_X86_GP_REG_UNDEF, 0,
-          (int)(im * l_vlen * 4 * i_micro_kernel_config->datatype_size_in + i_xgemm_desc->c1) );
+          (int)((long long)im * l_vlen * 4 * i_micro_kernel_config->datatype_size_in + (long long)i_xgemm_desc->c1) );
     }
 
     libxsmm_generator_gemm_decompress_i4_vreg ( io_generated_code, i_micro_kernel_config, i_xgemm_desc, l_zpt_vreg, l_vreg_lo, l_vreg_hi );
@@ -189,12 +189,12 @@ LIBXSMM_API_INTERN void libxsmm_generator_gemm_decompress_KxM_i4_tensor( libxsmm
         i_micro_kernel_config->vector_name,
         l_vreg_hi, (im == i_m_tiles-1) ? i_micro_kernel_config->mask_m_fp32 : 0, 0, 1 );
   }
-  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, i_gp_reg, i_ldi * 4 * i_micro_kernel_config->datatype_size_in );
-  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, o_gp_reg, i_ldo * 8 * i_micro_kernel_config->datatype_size_in );
+  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, i_gp_reg, (long long)i_ldi * 4 * i_micro_kernel_config->datatype_size_in );
+  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, o_gp_reg, (long long)i_ldo * 8 * i_micro_kernel_config->datatype_size_in );
   libxsmm_generator_gemm_footer_dequant_loop_amx( io_generated_code, io_loop_label_tracker, i_micro_kernel_config, cnt_reg, i_K/8);
   libxsmm_x86_instruction_pop_reg( io_generated_code, cnt_reg );
-  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_sub_instruction, i_gp_reg, i_ldi * 4 * (i_K/8) * i_micro_kernel_config->datatype_size_in );
-  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_sub_instruction, o_gp_reg, i_ldo * 8 * (i_K/8) * i_micro_kernel_config->datatype_size_in );
+  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_sub_instruction, i_gp_reg, (long long)i_ldi * 4 * (i_K/8) * i_micro_kernel_config->datatype_size_in );
+  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_sub_instruction, o_gp_reg, (long long)i_ldo * 8 * (i_K/8) * i_micro_kernel_config->datatype_size_in );
 #endif
 }
 
@@ -2832,7 +2832,7 @@ void libxsmm_generator_gemm_amx_kernel_mloop( libxsmm_generated_code*           
                 libxsmm_generator_brgemm_amx_set_gp_reg_a( io_generated_code, i_gp_reg_mapping, i_micro_kernel_config, i_xgemm_desc, l_src_a_id );
                 libxsmm_generator_gemm_getval_stack_var( io_generated_code, i_micro_kernel_config, LIBXSMM_GEMM_STACK_VAR_ELT_BUF1, l_gp_reg_scratch );
                 if (l_dst_a_id > 0) {
-                  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, l_gp_reg_scratch, l_dst_a_id*i_xgemm_desc->k*l_m_blocking);
+                  libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, l_gp_reg_scratch, (long long)l_dst_a_id*i_xgemm_desc->k*l_m_blocking);
                 }
                 libxsmm_generator_gemm_decompress_KxM_i4_tensor( io_generated_code, io_loop_label_tracker, i_micro_kernel_config, i_xgemm_desc,
                       m_blocking_info[l_m_count].tiles, i_xgemm_desc->k, i_xgemm_desc->lda, l_m_blocking,
@@ -2840,7 +2840,7 @@ void libxsmm_generator_gemm_amx_kernel_mloop( libxsmm_generated_code*           
               }
               libxsmm_generator_gemm_getval_stack_var( io_generated_code, i_micro_kernel_config, LIBXSMM_GEMM_STACK_VAR_ELT_BUF1, l_gp_reg_scratch );
               if (l_use_a_id > 0) {
-                libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, l_gp_reg_scratch, l_use_a_id*i_xgemm_desc->k*l_m_blocking);
+                libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_add_instruction, l_gp_reg_scratch, (long long)l_use_a_id*i_xgemm_desc->k*l_m_blocking);
               }
               libxsmm_x86_instruction_alu_reg( io_generated_code, i_micro_kernel_config->alu_mov_instruction, l_gp_reg_scratch, i_gp_reg_mapping->gp_reg_a);
               libxsmm_x86_instruction_alu_imm( io_generated_code, i_micro_kernel_config->alu_mov_instruction, i_gp_reg_mapping->gp_reg_lda, l_m_blocking);
